Create first simple class and test. Move prototype closer to where it should be implemented
